finished cart related work need to make add to cart behaveor on kaushalyas part
